posting to chatgpt api
