+++ server/server.ts
@@ -9,7 +9,7 @@ const PORT = 3000;
 /* 
 Import routers
 */
-import udemyRouter from './routes/udemyApi';
+// import udemyRouter from './routes/udemyApi';
 import chatRouter from './routes/chatApi';
 
 /*
@@ -33,7 +33,7 @@ app.get('/', (req, res) => {
 /* 
 Route handlers
 */
-app.use('/udemy', udemyRouter);
+// app.use('/udemy', udemyRouter);
 app.use('/chatgpt', chatRouter);
 
 /*
